add check for OPTR_LEGACY_MODE
There is a temporal relationship with flash-loop and the microcontroller which results in bricked MCUs. I think it is because power is interrupted or the microcontroller is reset when writing to OPTR. The frequency of bricking has increased because we have changed from a logical and with OPTR_MODE to an equality check for OPTR_MODE1 (0x1). The default value is OPTR_MODE_LEGACY (0x3), causing OPTR writes now on every flash...

A safer process would be to remain in OPTR_LEGACY, as this is compatible with OPTR_MODE1 behaviour.